move args loading and setting into ArgsToConfigUtils
